+++ runner/src/server/plugins/engine/pluginHandlers/exit/prehandlers/getBacklink.ts
@@ -1,0 +1,10 @@
+import { HapiRequest, HapiResponseToolkit } from "server/types";
+
+export async function getBacklink(
+  request: HapiRequest,
+  _h: HapiResponseToolkit
+) {
+  const state = request.pre.state;
+  const { progress } = state;
+  return progress?.at?.(-1) ?? null;
+}
