+++ runner/src/server/plugins/engine/pluginHandlers/exit/prehandlers/getState.ts
@@ -1,0 +1,6 @@
+import { HapiRequest, HapiResponseToolkit } from "server/types";
+
+export async function getState(request: HapiRequest, _h: HapiResponseToolkit) {
+  const { cacheService } = request.services([]);
+  return cacheService.getState(request);
+}
