+++ runner/src/server/plugins/engine/pluginHandlers/exit/prehandlers/parseErrors.ts
@@ -1,0 +1,7 @@
+import { HapiRequest, HapiResponseToolkit } from "server/types";
+import { errorListFromValidationResult } from "./utils";
+
+export function parseErrors(request: HapiRequest, _h: HapiResponseToolkit) {
+  const errors = request.yar.flash("exitEmailError");
+  return errorListFromValidationResult(errors);
+}
